requirements.txt
pytest
pytest-json
pytest-html
pytest-logger
pytest-selenium
python-i18n[YAML]

pyvirtualdisplay
